+++ apps/prairielearn/src/ee/auth/prairietest.html.ts
@@ -6,7 +6,7 @@ export const AuthPrairieTest = ({ jwt, prairieTestCallback, resLocals }) => {
     <!doctype html>
     <html lang="en">
       <head>
-        ${renderEjs(__filename, "<%- include('../../pages/partials/head') %>", resLocals)}
+        ${renderEjs(import.meta.url, "<%- include('../../pages/partials/head') %>", resLocals)}
         <style>
           .continue-card-container {
             width: 100%;