+++ apps/prairielearn/src/middlewares/selectAndAuthzAssessmentQuestion.js
@@ -1,10 +1,10 @@
 // @ts-check
-const asyncHandler = require('express-async-handler');
-import * as _ from 'lodash';
+import asyncHandler from 'express-async-handler';
+import _ from 'lodash';
 import * as sqldb from '@prairielearn/postgres';
 import { HttpStatusError } from '@prairielearn/error';
 
-var sql = sqldb.loadSqlEquiv(__filename);
+var sql = sqldb.loadSqlEquiv(import.meta.url);
 
 export default asyncHandler(async (req, res, next) => {
   const result = await sqldb.queryAsync(sql.select_and_auth, {