+++ apps/prairielearn/src/models/variant.ts
@@ -1,7 +1,7 @@
 import { loadSqlEquiv, queryAsync, queryOptionalRow } from '@prairielearn/postgres';
-import { Variant, VariantSchema } from '../lib/db-types';
+import { Variant, VariantSchema } from '../lib/db-types.js';
 
-const sql = loadSqlEquiv(__filename);
+const sql = loadSqlEquiv(import.meta.url);
 
 export async function selectVariantById(variant_id: string): Promise<Variant | null> {
   return queryOptionalRow(sql.select_variant_by_id, { variant_id }, VariantSchema);