+++ apps/prairielearn/src/pages/instructorAssessments/instructorAssessments.js
@@ -1,23 +1,23 @@
 // @ts-check
 import * as express from 'express';
-const asyncHandler = require('express-async-handler');
-import AnsiUp from 'ansi_up';
+import asyncHandler from 'express-async-handler';
+import { AnsiUp } from 'ansi_up';
 import * as error from '@prairielearn/error';
 import debugfn from 'debug';
 import { stringifyStream } from '@prairielearn/csv';
 import * as sqldb from '@prairielearn/postgres';
 import { pipeline } from 'node:stream/promises';
 
-import { courseInstanceFilenamePrefix } from '../../lib/sanitize-name';
-import { AssessmentAddEditor } from '../../lib/editors';
+import { courseInstanceFilenamePrefix } from '../../lib/sanitize-name.js';
+import { AssessmentAddEditor } from '../../lib/editors.js';
 import {
   updateAssessmentStatistics,
   updateAssessmentStatisticsForCourseInstance,
-} from '../../lib/assessment';
+} from '../../lib/assessment.js';
 
 const router = express.Router();
 const ansiUp = new AnsiUp();
-const sql = sqldb.loadSqlEquiv(__filename);
+const sql = sqldb.loadSqlEquiv(import.meta.url);
 const debug = debugfn('prairielearn:instructorAssessments');
 
 const csvFilename = (locals) => {
@@ -49,7 +49,7 @@ router.get(
       .filter((row) => row.needs_statistics_update)
       .map((row) => row.id);
 
-    res.render(__filename.replace(/\.js$/, '.ejs'), res.locals);
+    res.render(import.meta.filename.replace(/\.js$/, '.ejs'), res.locals);
   }),
 );
 
@@ -76,7 +76,7 @@ router.get(
     }
     res.locals.row = result.rows[0];
 
-    res.render(`${__dirname}/assessmentStats.ejs`, res.locals);
+    res.render(`${import.meta.dirname}/assessmentStats.ejs`, res.locals);
   }),
 );