+++ apps/prairielearn/src/pages/instructorCourseAdminSharing/instructorCourseAdminSharing.ts
@@ -1,13 +1,13 @@
 import { Router } from 'express';
-import asyncHandler = require('express-async-handler');
+import asyncHandler from 'express-async-handler';
 import * as error from '@prairielearn/error';
-import { InstructorSharing } from './instructorCourseAdminSharing.html';
+import { InstructorSharing } from './instructorCourseAdminSharing.html.js';
 import { z } from 'zod';
 import * as sqldb from '@prairielearn/postgres';
-import { getCanonicalHost } from '../../lib/url';
+import { getCanonicalHost } from '../../lib/url.js';
 
 const router = Router();
-const sql = sqldb.loadSqlEquiv(__filename);
+const sql = sqldb.loadSqlEquiv(import.meta.url);
 
 router.get(
   '/',