+++ packages/docker-utils/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @prairielearn/docker-utils
 
+## 2.0.11
+
+### Patch Changes
+
+- Updated dependencies [4f30b7e]
+  - @prairielearn/logger@2.0.0
+
 ## 2.0.10
 
 ### Patch Changes